+++ source/containers/Layout.js
@@ -58,7 +58,7 @@ export default class Layout extends Component {
 							<Route path="/contact" component={Contact} />
 							<Route path="/regle/:name" component={Rule} />
 							<Route path="/regles" component={RulesList} />
-							<Route path="/simu/:targets/:firstInput?" component={Simulateur} />
+							<Route path="/simu/:targets" component={Simulateur} />
 							<Redirect from="/simu/:name/intro" to="/simu/:name" />
 							<Route component={Route404} />
 						</Switch>
